Change the negation implementation to "Scheme Macro" (Transformer)
Instead of defining negation as a procedure, and invoke "take" inside it.
We used "Scheme Macro" to defining negation as a transformer.

In this way, we are making sure the completeness of the solving process.

Also, it makes the program more clear about "defining rules" and "solving/running". The negation is constructing an execution stream as mini-Kanren did before.