Added co-induction, simplified "no"/"complement rule" definitions.
Merged not_== to ==, on odd negation, we want =\=, a.k.a inequality. On even negation, we want == (original equality definition).

The calling stack frame (F) is tracking a record <Key, Value>. So the "element-of-set?" will return a record.

Also made changes to record representation (make-record, get-key, get-value)

During the goal expanding process, we inspected the calling stack frame (F) to see any ancestor call exist, and we break the infinite loop.

If we can't find an ancestor call, we will expand the calling stack based on "even/odd" negations. To invoke the "original" process or the "complement" process. (defined with prefix co_)

Therefore we simplified "def-asp-complement-rule" and current the transformation is done by hand.

Eventually, we simplified the "no" macro as well. We only need to add the negation counter (N) each time we see a negation. The predicate will decide to invoke the original form or the co-inductive form.