Add tracking and global constraint checking
Unlike positive program, the normal program should be treated as a whole part during the proving process, that is ALL rules must be executed at least once.

Therefore, we added a tracking set to track all ASP rules given to the system. Before the final-step "reify" we go over the rules that remain untouched.

If any of the untouched rule violates the unavoidable global constraint, a.k.a trigger the Odd Loop Over Negation (co-inductive failure). All checking processes will terminate, hence we got no result.

Otherwise the global constraint is avoidable, which means some checking process leads to success (no matter it's the original goal or the complement goal).