Fix the bug in forall implementation.
During the forall computation, if an intermediate variable (introduced by "fresh") get bind after a predicate evaluation.

We need to take all values (domain) of the variable, and check them with the future predicates.

Therefore, the execution of the future predicates is changing on the fly. If no variable is bound, it runs as (fresh-t (g ...)).
Otherwise, we need to use "take" to perform an eager execution to get the domain for variables. Then we are substituting var with each value and run (fresh-t (g ...)) with this substitution.

To succeed, the bounded variable need to succeed with all possible values, any failure will trigger a (mzero). Hence, we are using (bind *) to conjunct all results.

ToDo: we demonstrated with single variable for now, will support multiple variables later.