Passing original goal during checking.
During global constraint checking, we are making sure each value is provable. If the value can't be true or false, then it's unsatisfiable.
If a value failed, it could be failed by the OLON rules, or failed by other positive goals. The later one is a way to avoid the OLON rule. If a value failed by both OLON and positive goals, then the OLON is unavoidable. (Unsatisfiable).