Add grounding to the variable.
During NMR checking, we need to check each value for the rule.

We are using the idea of "reduct" to ground a predicate variable. Replacing all negation with a "top" (succeed). Then the program is safe to "take" all values.

The purpose of doing NMR-Check is to make sure each value has an external support. If any value reached an unfounded set, the program is unsatisfiable.