Beautify the code.
Line 304 getting the domain of a variable is cram in one long line, break it to multiple few lines, also we don't need to call "last-step" (perform NMR checking). We only do NMR checking at the end.

Line 378 Also here we are not running on the reduct program to compute the domain, it has to be the original program either in the positive form or negative form.

Line 395 the final result should be the accumulated info during the checking.

Line 419 remove unnecessary inc, we only have one entrance now, no inc and mplus* are needed.

Line 555 use exact match, since we use n == -2 for other purposes now.

Line 580 we don't update partial answer set during the grounding/NMR checking process.