Wrap up old project files
This project started in the summer of 2020 as a motivation for reimplementing s(ASP), but it turns out the s(ASP) method does not work. Luckily, the paper submitted to miniKanren 2021 and PPDP 2022 was not accepted. This project stopped in the fall of 2022, and I developed my approach to extend resolution and unification. This part of the new code in this repository I didn't commit, and I started a new project called stableKanren (https://github.com/stable-Kanren/stable-Kanren). I just committed everything on my local to clean up this project.